Restructure, update gym API
