+++ tests/test_gym.py
@@ -1,6 +1,8 @@
 import gymnax
-from gymnax.environments.gym import GymnaxToGymWrapper, GymnaxToVectorGymWrapper
+from gymnax.environments.conversions.gym import GymnaxToGymWrapper, GymnaxToVectorGymWrapper
+from gymnax.environments.wrappers.gym import RecordEpisodeStatistics
 import jax
+import jax.numpy as jnp
 import chex
 import gym
 
@@ -9,7 +11,7 @@ def test_gym_wrapper():
     env, env_params = gymnax.make("FourRooms-misc")
     e = GymnaxToGymWrapper(env, env_params, 0)
     a_space, o_space = e.action_space, e.observation_space
-    o = e.reset()
+    o, _ = e.reset()
     next_keys = jax.random.split(jax.random.PRNGKey(0), 2)
     o_env, env_state = env.reset(next_keys[1], env_params)
     chex.assert_trees_all_close(o, o_env)
@@ -34,10 +36,12 @@ def test_gym_vector_wrapper():
     assert isinstance(single_o_space, gym.spaces.Box)
     keys = jax.random.split(jax.random.PRNGKey(0), B)
     chex.assert_trees_all_close(e.rng, keys)
-    o = e.reset()
+    o, _ = e.reset()
     env_reset = jax.jit(jax.vmap(env.reset, in_axes=(0, None)))
     env_step = jax.jit(jax.vmap(env.step, in_axes=(0, 0, 0, None)))
     o_env, env_state = env_reset(keys, env_params)
     chex.assert_trees_all_equal_shapes(o_env, o)
     o, r, d, truncated, info = e.step(e.action_space.sample())
     e.render()
+
+
